+++ src/util.rs
@@ -3,7 +3,7 @@ use rand::Rng;
 /// Generate a random number between 150 and 300
 /// 
 /// It is used in hearbeat and election timeout
-pub fn random_timeout() -> i32 {
+pub fn random_timeout() -> u64 {
     let mut rng = rand::thread_rng();
     rng.gen_range(150, 301)
 }
